[INTERNAL] Move original serve-index code into serveIndex directory
